classifier/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
psycopg2-binary==2.9.9
pydantic==1.10.7
transformers==4.31.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.3.1+cpu
